# cmd/swag/docs/swagger.yaml
definitions:
  model.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.RegisterInput:
    properties:
      verify_key:
        type: string
      verify_purpose:
        type: string
      verify_type:
        type: integer
    required:
    - verify_key
    - verify_purpose
    - verify_type
    type: object
  model.SetupTwoFactorAuthInput:
    properties:
      two_factor_auth_type:
        description: 1:"email" and 2:"sms"
        type: integer
      two_factor_email:
        description: required if TwoFactorAuthType is "email"
        type: string
      user_id:
        type: integer
    required:
    - two_factor_auth_type
    - user_id
    type: object
  model.UpdatePasswordRegisterInput:
    properties:
      password:
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
  model.VerifyInput:
    properties:
      verify_code:
        type: string
      verify_key:
        type: string
    required:
    - verify_code
    - verify_key
    type: object
  response.ErrorResponseData:
    properties:
      code:
        description: status code
        type: integer
      data:
        description: du lieu tra ve
      message:
        description: thong bao loi
        type: string
    type: object
info:
  contact: {}
paths:
  /user/login:
    post:
      consumes:
      - application/json
      description: Login user and return authentication token
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: User login with credentials
      tags:
      - accounts user
  /user/register:
    post:
      consumes:
      - application/json
      description: Register a new user with OTP
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Register user with OTP
      tags:
      - accounts user
  /user/setup-2fa:
    post:
      consumes:
      - application/json
      description: Enables two-factor authentication for the user account
      parameters:
      - description: Bearer{token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.SetupTwoFactorAuthInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Setup two-factor authentication for user
      tags:
      - accounts 2fa
  /user/update-password-register:
    post:
      consumes:
      - application/json
      description: Allows a user to set or update their password after registration
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePasswordRegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Update password after registration
      tags:
      - accounts user
  /user/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify the OTP sent to the user
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.VerifyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      summary: Verify OTP for user
      tags:
      - accounts user
swagger: "2.0"